requirements.txt
marshmallow
pydantic
attrs
